Lots of readability-only cleanup.
Define a trivial __ne__() method in class Entity, causing most (all?)
other Cerebrum classes that need such a method to inherit one.

In general, it's rather retro to use a string as the first argument to
`raise' -- use class names instead.  If the string is something like
"Not implemented" it's much better to use Python's standard exception
`NotImplementedError'.