New and improved version.
TableView now supports (and requires) named columns. Data
can be added with both positional and keyword arguments.

A view can be changed to display certain columns.

A view can be sorted by any columns, shown or hidden, in reverse
or normal order.

A view can be returned as a twodimensional list or a HTML representation.