reverted previous patch which was bogous.
Please do not commit patches that "fix this very important problem
that I have _now_ in a way that I know is not correct".  If temporary
hacks like this are needed, checkout you own repository, and do the
changes there.  By checking them in, there is a risk that the hack
will be forgotten.