entity_info has support for OUs and Persons
entity_info doesn't return 'names' anymore. The 'names' was actually
never used, and it turned out to NOT WORK for OU - as some guy has
decided to implement a totally different get_names() for OUs than for
the rest of the entity types.

(And no - "SELECT * FROM ..." and just returning the rows without saying
what * is, is NOT a valid 'interface')


Two helper functions, str2num and num2str, makes you able
to write
  blapp = self.num2str(row.source_system)
instead of the more clumsy - always repeating one:
  blapp = str(self.num2const[int(row.source_system)])


I think this functionallity should be moved to Constants.ConstantsBase
so you might instead do:
  blapp = str(self.const.from_num(row.source_system))

or something like that.

This way - code other places than in bofhd_uio_cmds.py can do the same
trick instead of this str-int-munging.


entity_history didn't actually care about the limit of the number of
entries to be returned. Although no SQL-limitations are inforced in this
version, limits now ARE followed - although a dirty tricks was needed to
to unwrap the database Iterator to a normal list, and then slicing of
the limit last elements.