bugfix: user creation would cause "invalid disk" messages
todo: we are currently a bit inconsistent in the format of disks not
in the disks table.  _get_disk_or_home assumes a host:path syntax (we
don't store the host or use it for anything), while "user create"
expects a :path syntax.  The latter syntax cannot currently be used
with jbofh due to a bug in escaping it.