Check the encoding first. Uses sys.setenc() - a method that mirrors sys.setdefaultencoding()   -- to enable this, sitecustomize.py needs to add:      # save a copy for hacking since site.py deletes setdefaultencoding      sys.setenc = sys.setdefaultencoding
Sets default encoding to utf8.

That means that we still can use Cheetah (that does not support
unicode), and return stuff as UTF8.

However, this might be causing trouble when saving stuff over
xmlrpc, as xmlrpclib tries value = str(value) to see if it's ascii, instead
of doing value.encode("ascii").

(for example of encoding errors on submit, try to create a Note with
æøå)