Preparation for CVS <-> Arch synchronisation.
The sync script relies on a file whose name starts with "ChangeLog" to
get the log messages corresponding to the changes in the CVS
repository since the last sync attempt.

The plan is to autogenerate such a file from the CVS commit messages,
using "cvs2cl".  This autogenerated file must not be committed back
to CVS, as that would lead to an endless "empty" commit loop:

 * On the first run, the ChangeLog file at time t0 is autogenerated
   from CVS commit logs, and committed to CVS.
 * On the next run, the ChangeLog file is autogenerated again, this
   time including the initial CVS commit of the ChangeLog file.  This
   means there has been at least one non-CVS-committed change -- in
   the ChangeLog file.  The ChangeLog file is committed to CVS.
 * Repeat ad nauseum.

(Note that if our CVS repository already contained an
always-up-to-date ChangeLog file, the Arch sync script wouldn't need
to use cvs2cl to autogenerate it, and hence there would be no commit
loop.  However, as I have no idea on how we could enforce the
"always-up-to-date" criterion, I haven't pursued this variant any
further.)

As we might want to include (autogenerated) changelog files for
various Arch branches in the Arch repository, it seems like a good
idea to use a glob ("ChangeLog.*") for this CVS ignore pattern.