Revision: cerebrum@usit.uio.no--2004/Cerebrum--cvs--0--patch-218 Creator:  Harald Meland <hmeland@usit.uio.no>
New file: A (long) todo list towards making a 1.0 release.

The file is currently mostly written in Norwegian, and some parts of it
are probably weighted towards the needs we have at uio.no.

However, it will hopefully be more balanced later this week, after having
been discussed in the Cerebrum Developer Forum on Wednesday/Thursday.