Revision: cerebrum@usit.uio.no--2004/Cerebrum--cvs--0--patch-228 Creator:  Harald Meland <hmeland@usit.uio.no>
Adjustment to which SG packages we should build new Cyrus users on.

Some of the file systems we're running Cyrus on are nearly full; avoid
creating any more users on these.

Apart from the immediate problem of these disks filling up, this change
is also necessary to avoid any users being assigned to these servers
while they are being moved to the EVA (which will happen in the near
future).  Once this move is underway, further changes to this code will
be necessary.