Revision: cerebrum@usit.uio.no--2004/Cerebrum--cvs--0--patch-393 Creator:  Harald Meland <hmeland@usit.uio.no>
Make the ClassFronter sync of hia.no work after upgrade to FS 5.0

About a week ago, hia.no upgraded their student registry system, FS, to
version 5.0.  Some parts of the FS <-> Cerebrum interface was changed as
part of this update -- but no-one had thought it would affect the
ClassFronter export.  Until it turned out it did.

This fix will hopefully (re-)create Cerebrum groups for students in
various "kull".  The new groups have a slightly different name than they
used to, due to changes in FS 5.0.  The ClassFronter export, which relies
on these groups being up-to-date, has been slightly changed to deal with
the slightly changed names of the groups.