Revision: cerebrum@usit.uio.no--2004/Cerebrum--cvs--0--patch-413 Creator:  Harald Meland <hmeland@usit.uio.no>
FRIDA updates to conform with new XML schema.

The FRIDA project updated their import specification.  This patch
implements the necessary changes to produce an XML dump conforming to
this updated specification.

Note that this is a short-term solution, as it fetches its data from
other sources than Cerebrum db.  Until mod_hr is in place, it will have
to do.

As part of the patch, the generate_frida_export script now requires an
extra argument -- the path to the file "sted.xml".