Revision: cerebrum@usit.uio.no--2004/Cerebrum--cvs--0--patch-472 Creator:  Harald Meland <hmeland@usit.uio.no>
bofh_uio_cmds: Be more careful when comparing dates.

As mx.DateTime.DateTime objects (which is what pyPgSQL uses to represent
both DATE and TIMESTAMP values) allow comparison to None -- albeit with a
rather nonsensical outcome -- we need to be more careful whenever we
compare something to a database-retrieved date.