os.remove is more verbose than os.unlink (although they are the same function)
On win32, the os.rename() trick won't overwrite the existing file, so we
rename the existing file to .old first. Of course, then we also must
make sure that the .old file is deleted first.  If the main rename
operation fails, the .old file is rolled back to the original filename.

Note that this is no longer an atomic operation. Is this important? If
so, we should instead try to catch the rename "already exists" error and
do the .old trick in the exception catcher.