Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-66 Creator:  Harald Meland <hmeland@usit.uio.no>
list_disk_quotas: Improved handling of missing cmdline options

The usage message has been changed to reflect which options
are required.

Previously forgetting to specify either of the required options "-t
target_file" or "-s spread" would cause the script to crash with a Python
backtrace; now the usage message is presented, and the script exits with
a non-zero code.