***WARNING!
This is a database update. Key programs in Cerebrum may stop working
if you update the source without running the migration-script.
'contrib/migrate_cerebrum_database.py --help' for more details.

Most files should be updated, but errors may have sneaked by. Please
check your scripts if you think you may get errors from this.

NB! spine/ is not updated! spine-developers will have to update this
    themselves.



The reason behind this patch is to get ExternalIds on OUs. The code in
question is PersonExternalId which is moved and renamed to
EntityExternalId. The code should be portable enough to use in other
Entity-subclasses, if we wish. For now, Person and OU are the only
classes using it.

contrib/migrate_cerebrum_database.py is updated to ignore insert-codes
in pre. This means that you'll have to run 'makedb.py
--only-insert-codes' manually after migrating. (I figured local
maintainers wanted control over this step themselves.) The change is
done to allow renaming Constants in the API without makedb throwing
warnings and exceptions around. You should update codes after
migrating. Having different sets of Constants in the API and in the
database breaks stuff.

Fixes and usage of the new feature will follow.