Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-252 Creator:  Harald Meland <hmeland@usit.uio.no>
AccountUiOMixin: Use ChangeLog.change_by for BofhdRequest-requestor

If the Account's '_db' Database object is ChangeLog-enabled, and the
ChangeLog has a 'change_by' attribute, this attribute will hold the
entity_id of the account currently doing changes to the Account.  Use
this as 'requestor' of any automatically added BofhdRequests objects.

If there is no 'change_by' attribute, continue to use None as requestor.