Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-393 Creator:  Harald Meland <hmeland@usit.uio.no>
bofhd: Use the new Database.ping() method.

Bofhd wants to die if it's database connection no longer works, but the
old implementation didn't work with psycopg.  By using the
Database.ping() method instead, it should be easier to keep the check
working if support for more database drivers are added.