Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-395 Creator:  Harald Meland <hmeland@usit.uio.no>
Cursor.ping(): specify a column name to satisfy db_row.

Cursor.execute() uses the underlying cursor object's 'description'
attribute to build a db_row class corresponding to the column names in
the query.  The query in the ping() method caused this to fail, as it
didn't specify a db_row-acceptable column name when doing "SELECT 1".