GroupMember - get_groups is bugged, it returns groups who should be not included based on diffrence-membership. Not fixed, just commented.
EntityQuarantine - added a remove_quarantine, and made an important change to
add_quarantine. The old one would set the current date as start date nomather
which date was given as startdate, and null if no date was given. Instead it
now will set current-date if no date is given, or the date given.