Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-528 Creator:  Harald Meland <hmeland@usit.uio.no>
write_misc_info tweak

I do not like the fact that write_misc_info calls eval() on a value passed
to it directly from the command line. I think[tm] it makes more sense if
we select the proper bound method by using getattr -- we are likely to get
better error messages (AttributeError) and it is (albeit theoretically)
safer.