Readded support for remembering of which items was selected.
Added support for actions based on which items you select in the worklist.

TODO: firefox encodes urls in innerHTML, so the regexp-replacing is not
working, add action-html to the http-request sent when updating which
items was selected, load WL_actions from the html. And create
custom actions if several are selected.