Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-538 Creator:  Harald Meland <hmeland@usit.uio.no>
Added public interface for getting person_aff_status_code.status_str

The _PersonAffStatusCode class, which is our Python interface to the
person_aff_status_code database code value table, was missing a public
interface for getting the 'status_str' column of a code value.

Current code is using the internal interface ._get_status() for getting
this column; this ought to be changed.