Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-603 Creator:  Harald Meland <hmeland@usit.uio.no>
Fix buglet in AccountEmailMixin.update_email_addresses.

If the code changes the target type of an existing EmailTarget, it would
return once it's had updated the expire-dates of the target's addresses.

With this change, an "account"->"deleted" change will work as before,
while a "deleted"->"account" change will execute the rest of the method.