Changes to PERSON FIND
* Increased the minimum number of letters in the search pattern.  We
  are now no longer counting the wildcards % and _ as letters.
* Replaced the Exp-ID (which no one understands anyway) with
  primary account name (when available) in the output.  The export-id
  is still included in the reply, but it is not in the suggested
  output format.
* The name matching is more accurate, since it now specifies that
  matching should only be done against the cached full name.
  Previously, e.g. '%Kjetil' would return matches even though it's
  not a last name.

The lookup of the primary account is not free, but this is mostly
offset by the new API support for returning the name directly from
list_persons_by_name.  In fact, search for '%s_en' now runs in 24
seconds, but used 31 seconds in the old version.

This code requires a current version of Person.py