Adds two traits for default quota, one for hosts and one for disks. A disk *must* have the trait to enable quotas.
When a user is moved to a disk with more default quota (or no quota),
remove his quota (and tell operator about this in response to user_move.
This is done with a new mixin to Disk which overrides set_homedir.

DiskQuota.list_quotas now allows filtering by disk, and may include
accounts with no quota in the list.

New commands in bofh:
  host info
  host disk_quota
  disk list (replaces misc dls)
  disk quota
  email failure_message

  user move and user info have been updated to reflect the new
  quota system.