Revision: cerebrum@usit.uio.no--2005/Cerebrum--cvs--0--patch-838 Creator:  Harald Meland <hmeland@usit.uio.no>
CF fulldump: Don't export data for outdated courses

As the old-group-deletion code in populate_fronter_groups is currently
disabled, the Cerebrum database contains groups for outdated courses, and
some of these still have Fronter spreads.

This patch compares the set of groups that have fronter spreads with the
set of current courses in FS, and ignores the groups that correspond to
outdated courses.