Removed unecessary logger class.
Instead of having one Account instance (self.ac) we now have one in
the methods where it's needed.

All guest accounts now have the trait_guest_owner set, and when the
account is released the owner_id is set to None.

And some cleanup in logging messages, comments, etc is done.