"new" utility function spawn_and_log_output (moved in from process_bofhd_request) to run external processes and log the output.
scripts which use it should make sure to call get_logger() to
initialise the logger first, since this script relies on get_logger()
always returning the same logger object.

the caller should pass along the names of the servers which are
involved so that cereconf.DEBUG_HOSTLIST works according to
intention, but this is not required.

(it may be useful to add a companion function which captures stdout
and returns it.)