Added a proof of concept "Process Entity" script and API.
The goal
========
This is a working implementation of the 'automagic' that populates the
Cerebrum database with data not imported from authoritative systems. Data
like accounts, spreads, account affiliations and some group magic that
is needed by the Giske Indigo project.

Overview of the implementation
==============================

Current code
------------
The current implementation is a placeholder for a more dynamic set of
classes where config files will make up for a lot of the site spesific
code found in process_entity. The idea is that a set of basic tasks will
be implemented and controlled by some form of config directive. It is
implemented with a simple Factory.get() method to allow Mixins to override
current rules. The config file is currently only a few Mixin variables and
is in Python. The final language and syntax for this config file is TBD.

Both event based and batch handling
-----------------------------------
process_entity also implements both batch-like execution and future event
based execution. Today's batch running will only simulate an event like it
is received from an event handler. The goal with this implementation is to
be able to replace slow, seldom run batch jobs with faster, more complex
listeners or event handlers. Most of the code in process_entity is handling
an unknown entity where only the entity_type is known. This creates a
little overhead when running it as a batch job, but makes the code
portable.

The future
==========
process_entity needs a lot of work to be able to work as a generic "Black
Box" processor. The design is decent as it is, but a lot of the really
gritty config details are missing. There are plans to use this in other
Cerebrum projects so some cleanup and extension are expected shortly.
It is not expected to take over as the processor for the larger Cerebrum
instances like UiO since the ruleset will be near impossible to implement.
It is expected to take over for most other smaller instances, since a
common source base is A Nice Thing To Have(tm).