Enable proper handling of Norwegian characters.
The problem is that we want to be as flexible as possible wrt the input
objects. The clients should be able to pass strings and unicode objects
containing the Norwegian characters (Ã¸Ã¦Ã¥) without having to think about the
encoding and what not. This should Just Work[tm]:

logger.debug("blÃ¥bÃ¦rsyltetÃ¸y")
logger.debug(u"blÃ¥vÃ¦rsyltetÃ¸y")

... *regardless* of which encoding is configured for the logger.

codecs.StreamReaderWriter (the object returned from codecs.open()) is quite
nice, but it assumes that whatever is given to it is already in the proper
encoding. Effectively, it would force our client code to care about the
encoding of the objects passed along to the logger. This is a big usability
no-no. Thus, the encoding happens in the following way:

* All strings are forced to unicode objects[1]
* After that, all (unicode) objects are converted to the encoding specified
  in logging.ini (before being output). If no encoding is specified, latin1
  is the default.

[1] There is probably a penalty associated with all this conversion magic.
Someone should figure out how much this costs us.