Fixing inconsistencies.
Up until now it was possible for a person to be part of kull/und.enhet/etc.
groups in the file, but be otherwise missing (i.e. no <person> element). This
is unfortunate, but it was possible, since the sources for these elements are
different (FS and Cerebrum).

As of this patch, it is no longer the case. Before an fnr from FS is output, it
is re-mapped to the fnr in Cerebrum. This remapping will fail, if a
corresponding <person>-element has not been output earlier.