Added skeleton confirm-action stuff for non-javascript browsers.  It's simple but it should work.
Now you can prepend an url with /confirm/ and the user will be told the action is final and asked whether he wants to do it or not.

So /account/delete should now be /confirm/account/delete.  Next up is making javascript to intercept /confirm/* and asking the user in a popup.