Changes due to lt -> sap
In general lt is not removed, but sap is just added to the list of
source_systems. Quite a few places there was logic that was nice and
simple for to source_systems (fs_lt) but when sap is put into the
equiation there code needed some extension. Hopefully thing will not
break because of this. Testing is needed.