Hopefully this fixes exceptions in spine for good.
Make unhandeled exeptions more like handled exceptions,
except that a stacktrace is dumped to the log and all such errors gets
mapped to ServerError. (But we retain the name and message of the
original error in the description.)

Then cut some errors from the list of handeled errors, these were really
serverside, and should genereate a stack trace. The new better handling
of unhandeled exeptions is sufficient to categorize bugs.

This should provide both data for debugging serverside, and still give
the user a hint of what happened. (=> Much better error reports.)

UnknownException should never occur after this!