SAP data files refer to OUs that do not exist in SAP but that do exist in LT. This creates a host of problems, so we'll try to prevent that.
The idea is NOT to assign affiliations, unless an OU for the affiliation exists in the 
proper hierarchy (wrt source system for the import). This way we make sure that although
the sko that is referred to exists in cerebrum, it will be ignored, if it is not registered
in the same source system as the import data are coming from.