Fix for concurrency issues arising from rotating handlers.
The problem lies in simulnateous access to the same logger from
multiple threads. Under certain conditions (as always, this is a
question of timing), two threads may decide to rollover the same log
file simultaneously. This leads to termination of one of the threads
and potentially loss of data. All in all, an obviously undesirable
situation. 

However, equipping a handler with a lock will give us a possibility of
doing the rollover within a critical region. Hopefully I got the
nitty-gritty details right on this one. 

Caveat: if two different handlers work on the same file, *or* each
thread gets its own logger (pointing to the same file), there is no
way we can prevent chaos.