Quell the error stream about nameless OUs.
next_object() decides how a suitable object is constructed from the
next XML element. If no such object can be created (because the data
is foobar), return None. This way, when an OU or a Person cannot be
created for some known/expected reason, we can communicate it without
resorting to exceptions, since the latter are somewhat tricky to deal
with in the logs.