Entity Email patch. This will break your installation horribly if you don't upgrade to mod_email_1_3.
Most code is patched to email_1_3, but Spine stuff is left in
peace. NTNU has to fix this. 

This upgrade is large. You should test an upgrade on a test-server
just to verify that it will go according to plan. 

=====================================================================
IMPORTANT!!! After upgrading you should run "vacuum analyze;" in psql
due to indexes being dropped and re-created. If not, your performance
will be poor.
=====================================================================

A lot of code is tested, but not all. Expect some of it to break. Test
your scripts and services before you migrate.