Logging and misc.
-make client send hostname and client version
-log loging/logouts
-change debug_transactions to return the result to the client
-fix (again) autologout from SpineClient

The spine (internal) login-method is changed to login2, and now
needs more arguments. These are filled inn automatically by the client.
(Users still use a compatible login method)

The SpineCore client API changes with this change. Clients and server
need to be updated at the same time. It may be necessary to remove
temporary files for some clients.

While it is certainly possible to prossible to provide a login
method for old clients, the goal of this change is to have a framework
to force clients to upgrade. A non-versioned login method would defeat
this purpose.