It is non-trivial to reconcile old- and new-style classes across python 2.4 and 2.5. Hope I have it right this time.
Essentially, the problem boils down to the right way of monkey patching our
own Error/Warning classes into the db-backend specific Error/Warning classes
in such a way that it works both with 2.4 and 2.5. Additionally, the built-in
exception classes themselves changed between 2.4 and 2.5.

A sensible test to run in the future is:

>>> import cerebrum_path, cereconf
>>> from Cerebrum.Utils import Factory
>>> db = Factory.get("Database")()
>>> g = Factory.get("Group")(db)
>>> import Cerebrum.Database
>>> issubclass(g._db.IntegrityError, Cerebrum.Database.Error)
True
>>> try:
...   raise g._db.IntegrityError("fdjas")
... except Cerebrum.Database.Error:
...   print "got it"
... 
got it
>>> 

This must work regardless of the python version.