This is getting more aggravating by the commit...
Devs, beware -- as of python 2.5, you cannot use types in "except:" that do
not inherit from BaseException. 

Fixed Database.py so that our classes are monkey patched properly. This should
work with 2.4 and 2.5 now. Added a few more tests (not in this commit). There
is still the issue of MRO -- since the backends are written in C, it is not
always possible to rearrange __bases__. A thing to consider is to reassign
__str__, so that our __str__ takes precedence. The best suggestion I have thus
far is still too hairy for common consumption.