Please vortex-core.
This patch creates a number of additional eduPersonEntitlement attributes in
LDAP. Vortex group has requested to have access to FS roles assigned to people
in undenh/undakt (EVU-kurs/kursakt/kull are not quite that important right
now). There used to be an 'enhentsansvarlig'-like entitlement, which
disappeared when DML requested changes to fronter roles. However, vortex
people still need that information (e.g. in order to give permissions for
course websites to those who are somehow involved with the courses).

So, this is how it works now:

1) Someone in FS assigned a role (like 'dlo', 'gruppelære' or 'it-ansvarl') to
   a person P. Let's say role 'DLO' to P for undenh U.
2) Cerebrum fetches that information from FS and registers this in roles.xml
3) populate_fronter_groups.py reads roles.xml and creates a group in cerebrum
   where P is a member. This group is called something like
   'uio.no:fs:<enhetid>:<rolletype>', which in this case is
   'uio.no:fs:<U>:dlo'.
4) generate_kurs_ldif.py generates a list of undenh/undakt "entities"
   available for LDAP. Additionally, generate_kurs_ldif.py registers which
   people have which entitlements with regard to such undenh/undakt. The
   information about entitlements is stored in ownerid2person pickle file. 
5) generate_org_ldif.py loads the said pickle-file and generates the
   corresponding eduPersonEntitlement entries in the org.ldif file. For our
   guinea pig P this will look something like:

   eduPersonEntitlement: DLO@urn:mace:uio.no:section:aktivitet-888712_888713

   ... where the urn designates the undenh U.

Now this attribute is available in LDAP. Vortex people can access it as they
do with the "Learner@<urn>" attributes. The only thing they need to know is
what kind of attributes to expect before the urn. A separate e-mail on the
topic will be sent.

The entire construct is somewhat involved and partly hackish. If we get
additional roles or (flying spaghetti monster forbid) additional "entities"
where the roles are relevant (say "kull" og "EVU-kurs"), the plan sketched
above would need a rework (generate_kurs_ldif.py is not capable of supporting
kull og evu-kurs right now).