There was a subtle error here. If a user has spread to more than one domains and the ad attributes should be set only one of them was set.
The problem was that we used the dict self.entity_id2profile_path (and
for ou and homedir), which is created at the start of the program, to
get the spread2profile_path (and vice versa for ou and homedir). The
dict cannot be used as a base to populate a trait in populate_ad_attrs
because the users spread2profile_path may have been updated during the
execution of the script. Hence we have to use the API to get the trait.

 --This line, and those below, will be
ignored--

M    process_ad_spread.py