Wofh needs to know what groups a given operator can administer/moderate.
This method (access_list_alternable) is a slight generalisation, and allows
operators to retrieve all the groups, disks hosts and maildoms that they have
access to. This is a complement/generalisation of the can_alter_group
permission (we retrieved all groups that a given operator has can_alter_group
permission on). 

Caveat: For superusers the result is somewhat "broken" (we return only the
entities where a superuser operator would have had permissions, had (s)he not
been superuser).

Caveat: There is no cutoff for the number of entities returned (potentially we
could return the entire content of group_info).

Caveat: All users are allowed to run this command (on themselves, obviously).