Introduce support for non-advertised commands.
This is a good idea, but a hackish implementation.

Some commands should exist, but should not be advertised to clients. I.e. the
commands should be implemented in such a way as if any client knew about it
(there is no security in obscurity). However, such hidden commands will not be
published in e.g. jbofh (or similar clients requesting available commands via
the get_commands() framework), which makes hidden_commands an ideal hook for
administrative tasks.

BEWARE! hidden_commands MUST validate all of their arguments.