Small fix for perm list.
The problem arose around "perm list 42".

bofhd_uio_cmds.py gets 42 as a string, not as an int. While checking
permissions, BofhdAuthRole.list() gets a list containing one element, string
"42" and does this:

"%i" % "42"

which is an error in Python (whereas "%s" % "42" would not have been).

This patch converts the strings to proper values.