Felles Studentsystem (FS) is being upgraded to v6.2. In this version the eksamensmodul is replaced by the vurderingsmodul.
The functions regarding the tables in the eksamensmodul are therefore
updated accordingly. This applies to the following functions:

 * _list_aktiv_semreg
 * _list_aktiv_enkeltemne
 * _list_aktiv_avlagteksamen
 * list_privatist_emne
 * list_studenter_alle_undenh

Please note that this version of access_FS.py will not work with FS
versions below 6.2.

Also note that these new functions should not be put in production
without testing to ensure that the data is consistent after FS
is upgraded to v6.2 and converted.