Put the db-backend testing framework in place.
You wanna write a new test? Well, here you go:

* For backend-specific tests, add a test to test_<your backend>.py
* For generic db backend tests, add a test to common.py

You wanna write a test suite for a new backend?

* Create test_<backend>.py
* Create

  import common
  class test_whatever(common.DBTestBase):
    def setup(self): 
      # something
    def teardown(self):
      # something

Be mindful of what kind of external environment is required. The goal is to
require as little as possible, so that people would not need half a century to
set testing up.