Methods for manipulating AD attributes, which are stored as EntityTraits, are added. This is done to hide the trait ugliness such that client code is simplified.
Note that delete_spread is overridden, to make sure that AD attributes
are deleted for an account that looses spread to an AD domain. This is
a bug fix.

Also note that a __init__ method is added. Make sure that
CLASS_ACCOUNT in cereconf is correct before running the code in
production environment.