Added two new methods, group_multi_add and group_multi_remove, which does the same as all the variants of group_?add and group_?remove, with the first argument as the membertype (group, account or person). They are calling _group_add and _group_remove directly, so they're hopefully without that many bugs.
For now are these hidden commands, so jbofh-users don't get confused by a new command.
It is, however, the plan to make this the new group_add and group_remove.

Another thing that is added, is the possibility to remove members from a group by their
entity_id. Since group_list returns full name and entity_id of a person, this is the
easiest way to do it for brukerinfo. This may not be the best way of doing it, and others
may please look into it, since this is my first mess in the bofhd code :)