A few updates according to feedbacks (see cerebrum_sites/doc/intern/uio/utvikling/webgrensesnitt-jokim/tilbakemeldinger.rst).
Logon:
Fixed a bug which gave the user a connection error message if the username or password was wrong.


Person:
Contact information for both FS and SAP are now constantly shown, and not depending on where you are registered, as the registrations could be wrong.


Accounts:
Expired users are now marked as "$username (deleted)" in Your other accounts-list, together with the expired-date.


Email:
The list of other email accounts doesn't include epxired users anymore.


Groups:
Previously, if you could moderate a group, the group would be removed from the list of your group memberships. This is now fixed so groups could be in both lists.

Also fixed the message for creating a personal group, so it actually checks if the group exists.