Added two new functions to the xmlrpc-class, bofhd_to_native and native_to_bofhd, to handle :None and escaping other :, in addition to the character encoding (utf8 <-> iso-8859-1).
Retreiving strings from bofhd is done in utf8, but sending strings to bofhd is more bogus. This means that
data sent to bofhd is converted to iso-8859-1, while data from bofhd is not converted (except of the 
: escaping).


There is also implemented a locking function. The site checks if the file
  data/lock
is not empty, it will log out every user and only give them a page with the text from the lock-file.
If the lock-file exists, but is empty, nothing will happen.

An example-file is added:
  data/lock_default
This could be updated with a default text and using the lock as a link to it
  ln -s lock_default lock

An roughly update procedure would then be:
1. Create a link from lock to lock_default
2. Update the system files
3. Remove the lock-link

This would work for most updates, except if there are changes in cached data.