Implemented a en masse password setting feature requested by Giske.
Giske wants to be able to set passwords in bulk for an entire group in one
go (this makes sense when the want to re-set passwords for all pupils in a
specific class/group). This feature has been enabled for c3 users
(administrators). An admin has a choice of setting a new password for all
members of a group, or a selected bunch. A list of altered passwords is
reported back to the administrator for printout/whatnot.

Do NOT launch this patch into live environment without properly adjusting
cereconf.CWEB_AVAILABLE_COMMANDS.