Implemented new rules for respecting employment termination dates.
In essence, this patch respects termination dates (fratredelsesdato) in
feide_person.txt -- any employee with such a date set to some time in the past
will no longer be considered an employee and all of his/her employment
information will be ignored. An immediate consequence of this will be the loss
of all ANSATT-affiliations.

Do not install this into the production environment without a warning to the
respective institution.