Temporary hack to appease eh... someone.
The idea is that we need to populate a group NIS map based on groups that may
contain people members. This kind of sucks, since the actual members in the
NIS map are accounts, not people. This hack implements person -> primary
account mapping (for certain Cerebrum groups) while generating NIS maps.