Script to create/reactivate accounts for holders of specific affiliations.
This script is supposed to offset some of the manual work related to new
employees. The script scans the db for a specified affiliation (or a set
thereof) and checks that people holding that affiliation have a non-expired
account. Those who do not, get a new account with the same set of affiliation
as the owner.