On 2009-08-24 giske reported that they experienced intermittent failures related to simultaneous cweb usage. The problem has been traced down to improper state database access.
Cweb relies on gdbm to store session-related information. Under a number of circumstances 
cweb required write access to that db; and by default the state db is opened in "write 
access" mode. gdbm, however, does NOT support non-exclusive write access, which results in 
access failure, should two users attempt to open the state database at the same time.

This patch attempt to avoid, rather than actually correct the problem -- in case of access 
failure the application waits a little bit before attempting to open the state database 
again. This is NOT a proper solution, and gdbm usage must be deprecated in favour of a 
proper backend (such as postgres).