Deal with utf-16 in a sane fashion.
OEPAPRD changed the charset (or was it the table definisjon) from latin-1/utf-8 to utf-16. 
DCOracle2 driver, which interprets NCHAR/CHAR datatypes as strings, effectively
returns 2 byte sequences for the chars in the "username range" ([a-zA-Z0-9]). Naturally, 
dbfg_update.py does not cope well with it, since cerebrum content is fetched and 
represented in latin-1 internally. 

The proper solution is to treat everything as unicode objects and use a driver that does 
just that (cx_oracle). 

The slightly less demanding solution is to recode the data coming from OEPAPRD to always be 
in latin-1. Not pretty, but it works.