Giske wants to see affiliations in lists of members in a group. Finally they will get it.
The change will create a bit more load, as it will call person_info for every
account or person in the group. A list is limited at 300 members, so loading a
page will take a few more seconds. If we want to raise the limit of max members
we could create a new bofh-command which gathers data about each member
internally in bofhd to make it faster (e.g. 
group_list_with_member_affiliations()).

This will also affect ÃfK.