A bofh session tripped an error here.
list_external_ids() by default returns a generator object, which always
evaluates to True in a bool context. Since we want to check whether
there are any ids at all matching the criteria, it's easiest to do
so once the actual id list (or set, or tuple, but NOT a generator) has 
been created.