Force ping() NOT to alter the transaction state.
This is one insidious little fucker. Previous revision (r12892) made an
erroneous assumption that forcing ROLLBACK after ping was a safe course
of action given our ping() usage pattern. This is wrong, since makedb.py
manipulates the constants in a non-trivial way, and ping()-ing with a 
forced ROLLBACK would result in new constants being dropped (the error       
can easily be triggered by using makedb.py --only-insert-codes with a new 
constant) due to ROLLBACK: our Constants.py:_CerebrumCode does an 
automatic ping() before retrieving the database proxy it uses to insert 
or to fetch constant values. If ping()-ing rolls the transaction back, 
freshly inserted constants will be dropped as well.

Therefore, the sensible course of action is to make sure that ping()-ing 
happens as an atomic action that DOES NOT alter the state of the surrounding 
environment. This can be accomplished (albeit in a db-specific manner) by
using nested transactions.

This patch introduces support for our current backend of choice, psql. A 
ping is initiated within its own savepoint (and ONLY inside a cursor). The 
savepoint encompasses just the ping call and, therefore, will not affect 
the state of the surrounding transaction. 

Unfortunately there is still an issue of multiple savepoints from different 
cursors being started simultaneously on the same db connection. There is a 
possibility that one of the cursors will report a failure in ping(), because
another cursor rolled to a savepoint started earlier. However, this should 
not cause any problems in practice.